+++ src/components/sidebar/index.js
@@ -1,2 +1,1 @@
-export { default as Sidebar } from "./sidebar.component";
-export { default as SidebarHeader } from "./sidebar-header/sidebar-header.component";
+export { default } from "./sidebar.component";