+++ src/utils/helpers/options-helper/options-helper.js
@@ -267,6 +267,8 @@ const OptionsHelper = {
 
   sizesPod: ["extra-small", "small", "medium", "large", "extra-large"],
 
+  iconSizes: ["small", "medium", "large", "extra-large"],
+
   sizesRestricted: ["small", "medium", "large"],
 
   sizesBinary: ["small", "large"],